pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hatctl"
version = "2.1.0"
description = "Put on your company hat — switch between company environments"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Alexander Pyatkin", email = "av.pyatkin@gmail.com"},
]
readme = "README.md"
keywords = ["cli", "devops", "context-switching", "company", "environment"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "keyring>=25.0",
    "httpx>=0.27",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/apyatkin/personal-tools"
Repository = "https://github.com/apyatkin/personal-tools"

[project.scripts]
hat = "hat.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hat"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff.lint.per-file-ignores]
"src/hat/cli.py" = ["E402"]
"tests/*" = ["E402"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.8",
]
